+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/IrContributedGraphGenerator.kt
@@ -2,13 +2,13 @@
 // SPDX-License-Identifier: Apache-2.0
 package dev.zacsweers.metro.compiler.ir
 
+import dev.zacsweers.metro.compiler.NameAllocator
 import dev.zacsweers.metro.compiler.Origins
 import dev.zacsweers.metro.compiler.Symbols
 import dev.zacsweers.metro.compiler.asName
 import dev.zacsweers.metro.compiler.capitalizeUS
 import dev.zacsweers.metro.compiler.decapitalizeUS
 import dev.zacsweers.metro.compiler.exitProcessing
-import dev.zacsweers.metro.compiler.joinSimpleNames
 import org.jetbrains.kotlin.backend.jvm.codegen.AnnotationCodegen.Companion.annotationClass
 import org.jetbrains.kotlin.descriptors.ClassKind
 import org.jetbrains.kotlin.ir.UNDEFINED_OFFSET
@@ -27,7 +27,6 @@ import org.jetbrains.kotlin.ir.expressions.impl.IrInstanceInitializerCallImpl
 import org.jetbrains.kotlin.ir.types.IrSimpleType
 import org.jetbrains.kotlin.ir.util.addChild
 import org.jetbrains.kotlin.ir.util.addFakeOverrides
-import org.jetbrains.kotlin.ir.util.classIdOrFail
 import org.jetbrains.kotlin.ir.util.copyAnnotationsFrom
 import org.jetbrains.kotlin.ir.util.createThisReceiverParameter
 import org.jetbrains.kotlin.ir.util.defaultType
@@ -41,6 +40,8 @@ internal class IrContributedGraphGenerator(
   private val contributionData: IrContributionData,
   private val parentGraph: IrClass,
 ) : IrMetroContext by context {
+
+  private val nameAllocator = NameAllocator(mode = NameAllocator.Mode.COUNT)
 
   @OptIn(DelicateIrParameterIndexSetter::class)
   fun generateContributedGraph(
@@ -107,13 +108,8 @@ internal class IrContributedGraphGenerator(
         .buildClass {
           // Ensure a unique name
           name =
-            sourceGraph.classIdOrFail
-              .joinSimpleNames(separator = "", camelCase = true)
-              .asSingleFqName()
-              .pathSegments()
-              .joinToString(separator = "", prefix = "$\$Contributed") {
-                it.asString().capitalizeUS()
-              }
+            nameAllocator
+              .newName("$\$Contributed${sourceGraph.name.asString().capitalizeUS()}")
               .asName()
           origin = Origins.ContributedGraph
           kind = ClassKind.CLASS